583. Delete Operation for Two Strings
